chore: move connections store
